More cleaning up things and fixed the circular issues
