--- custom_components/sat/device.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class DeviceState(Enum):
-    ON = "on"
-    OFF = "off"
